Support use of the CMake build system in embedded projects.
Add control of library build type through the BUILD_SHARED_LIBS CMake variable. This is necessary for bare-metal targets that do not support shared libraries. The default behavior is still to build a shared library.

Enable project configuration when embedded as a CMake subdirectory. Checks if a variable is defined in the enclosing scope before adding a cache entry. This became the default for the option command in CMake 3.13. See CMake policy CMP0077 for more explanation.

Add SB_TEST CMake option to control building of tests. The default behavior is still to build tests.

Add public headers directory to library target interface. This allows consuming CMake targets to easily add the library as a dependency.